[#75113332] Create Mailchimp::List model

+++ db/migrate/20140718215456_remove_list_from_mailchimps.rb
@@ -1,9 +1,5 @@
 class RemoveListFromMailchimps < ActiveRecord::Migration
-  def up
-    remove_column :mailchimps, :list
-  end
-
-  def down
-    add_column :mailchimps, :list, :string
+  def change
+    remove_column :mailchimps, :list if table_exists?(:mailchimps)
   end
 end
