+++ db/migrate/20140723144855_drop_mailchimp_integrations_2.rb
@@ -1,6 +1,6 @@
 class DropMailchimpIntegrations2 < ActiveRecord::Migration
   def up
-    drop_table :mailchimp_integrations
+    drop_table :mailchimp_integrations if table_exists?(:mailchimp_integrations)
   end
 
   def down
