+++ db/migrate/20140723153033_drop_mailchimp.rb
@@ -1,6 +1,6 @@
 class DropMailchimp < ActiveRecord::Migration
   def up
-    drop_table :mailchimps
+    drop_table :mailchimps if table_exists?(:mailchimps)
   end
 
   def down
